+++ src/js/app.js
@@ -1,0 +1,40 @@
+/* ===================================
+
+	Main JS document for Application
+
+======================================*/
+
+// for this project, will use the MVC sample defined here : https://alexatnet.com/articles/model-view-controller-mvc-javascript
+
+
+	/* ===================================
+
+		Model
+
+	======================================*/
+
+		 function Model(controller){
+
+		};
+
+	/* ===================================
+
+		View
+
+	======================================*/
+
+		function View(model, controller, elements){
+
+		};
+
+	/* ===================================
+
+		Controller
+
+	======================================*/
+
+		function Controller(view, model){
+
+		};
+
+export default main;