+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ** Changelog for Event Planning Application
 
+v0.2
+- changed UI to fit project direction
+- added base MVC model for app.js and my-app.js
+- added polymer UI components to app, and bower.json
+- styled some UI elements to better fit project direction
+
 V0.1
 - base version created.
 - added dependency files, and gulp dependencies.